Partial changes in CRUD API Posts and Recipes

+++ src/main/java/com/sinter/cookspire/dto/CommentDTO.java
@@ -16,10 +16,10 @@ public class CommentDTO {
     private String content;
 
     @Nullable
-    private long likes;
+    private int likes;
 
     @Nullable
-    private long dislikes;
+    private int dislikes;
 
     @Nullable
     private LocalDateTime createdOn;
